refactor: add cta-slot to grant-item component
completes JOB-689
