+++ src/grants/components/grant-card/index.ts
@@ -1,0 +1,1 @@
+export * from './grant-card';
