+++ src/grants/components/grant-item/grant-item.stories.tsx
@@ -8,6 +8,7 @@ const meta: Meta<typeof GrantItem> = {
   component: GrantItem,
   args: {
     grant: fakeGrant,
+    cta: null,
   },
 };
