+++ src/grants/components/grant-list/cta.tsx
@@ -1,0 +1,31 @@
+import { Button } from '@nextui-org/react';
+
+const CaretRightIcon = () => (
+  <svg
+    width="32"
+    height="33"
+    viewBox="0 0 32 33"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M12.2929 10.7929C12.6834 10.4024 13.3166 10.4024 13.7071 10.7929L18.7071 15.7929C19.0976 16.1834 19.0976 16.8166 18.7071 17.2071L13.7071 22.2071C13.3166 22.5976 12.6834 22.5976 12.2929 22.2071C11.9024 21.8166 11.9024 21.1834 12.2929 20.7929L16.5858 16.5L12.2929 12.2071C11.9024 11.8166 11.9024 11.1834 12.2929 10.7929Z"
+      fill="#F9FAFB"
+    />
+  </svg>
+);
+
+export const GrantListCTA = () => {
+  return (
+    <div className="flex size-40 items-center justify-center">
+      <div className="flex items-center gap-4">
+        <Button color="primary">
+          <span>Apply</span>
+        </Button>
+        <CaretRightIcon />
+      </div>
+    </div>
+  );
+};
