+++ src/grants/components/grant-list/grant-list.tsx
@@ -1,6 +1,8 @@
+import { GrantItem } from '@/grants/components/grant-item';
+
 import { fakeGrant } from '@/grants/testutils/fake-grant';
 
-import { GrantItem } from './grant-item';
+import { GrantListCTA } from './cta';
 
 export const GrantList = () => {
   // TODO: JOB-682
@@ -8,7 +10,7 @@ export const GrantList = () => {
   return (
     <div className="flex flex-col gap-4">
       {Array.from({ length: 64 }).map((_, index) => (
-        <GrantItem key={index} grant={fakeGrant} />
+        <GrantItem key={index} grant={fakeGrant} cta={<GrantListCTA />} />
       ))}
     </div>
   );
