+++ src/grants/components/grant-list/index.ts
@@ -1,0 +1,1 @@
+export * from './grant-list';
